requirements.txt
python-pptx
pypdfium2
rich
questionary
flet
easyocr
iopaint
opencv-python
scikit-image
numpy
